feat: Move database back to aggregator
