+++ aggregator/database/gen.go
