+++ aggregator/database/mocks/database.go
